Refactor product management to include product type and unit
- Updated ProductController to handle product type and unit in create and update requests.
- Enhanced ProductDto, ProductCreateRequest, and ProductUpdateRequest to include product type and unit fields.
- Modified Product entity to store product type and unit, with default values for backward compatibility.
- Adjusted ProductService to normalize product type and unit during creation and updates.
- Updated Stock and SaleItem entities to use BigDecimal for quantity to improve precision.
- Implemented a backfill service to update existing products with default values for type and unit.
- Enhanced EntityMapper to map new fields into ProductDto.

+++ src/main/java/com/tamali_app_back/www/dto/StockDto.java
@@ -1,5 +1,6 @@
 package com.tamali_app_back.www.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record StockDto(UUID id, UUID productId, int quantity) {}
+public record StockDto(UUID id, UUID productId, BigDecimal quantity) {}
