+++ src/main/java/com/tamali_app_back/www/dto/StockMovementDto.java
@@ -2,6 +2,7 @@ package com.tamali_app_back.www.dto;
 
 import com.tamali_app_back.www.enums.MovementType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -11,7 +12,7 @@ public record StockMovementDto(
         UUID businessId,
         UUID userId,
         String userDisplayName,
-        int quantity,
+        BigDecimal quantity,
         MovementType type,
         LocalDateTime movementAt
 ) {
